Use last-modified date from GitHub for condition
On branch HEAD update, the last-modified is the date of the commit. So
we can have a cache newer than the commit, but the data is actually
older. The bug can also occur when we hit a race condition.

So we hack githubpy to get HTTP headers. We save headers in the cache
and do the conditionnal query with actual Last-Modified date.

This fixes pushes not being built because GHP is still checking the
previous commit.