Refactorize list_projects
The purpose of this changes is to let the CI server maintain the list of
managed projects. Since Jenkins1 does not managed project without jobs,
the JENKINS client complete Jenkins introspection with values from
settings.

Anyway, the CI server si also responsible to known which branches to
build. This si considered a CI settings, not in poller. So we add
branches listing right in project listing.