Fix URL for accessing the API of a build
At some point between Jenkins 2.52 and 2.59, the URL returned at job
creation by jenkins changed from the url of the job to that url suffixed
with /display/redirect. When suffixing all that with /api/python, the
api part is ignored. This commit removes the /display/redirect just
before using it.